# docs/swagger.yaml
definitions:
  domain.Order:
    properties:
      created_at:
        type: string
      customer_name:
        type: string
      id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
  domain.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /order_handler:
    post:
      consumes:
      - application/json
      description: Create a new order_handler
      parameters:
      - description: Order
        in: body
        name: order_handler
        required: true
        schema:
          $ref: '#/definitions/domain.Order'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Order'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new order_handler
      tags:
      - order_handler
  /order_handler/{id}:
    get:
      consumes:
      - application/json
      description: Get an order_handler by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Order'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get an order_handler by ID
      tags:
      - order_handler
    put:
      consumes:
      - application/json
      description: Update an order_handler by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order
        in: body
        name: order_handler
        required: true
        schema:
          $ref: '#/definitions/domain.Order'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Order'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update an order_handler by ID
      tags:
      - order_handler
  /users/login:
    post:
      consumes:
      - application/json
      description: Login a user_handler
      parameters:
      - description: Credentials
        in: body
        name: credentials
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login a user_handler
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Register a new user_handler
      parameters:
      - description: User
        in: body
        name: user_handler
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register a new user_handler
      tags:
      - users
swagger: "2.0"